pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cartrawler-mcp"
version = "0.1.0"
description = "CarTrawler MCP Server for ChatGPT - Flight, Car, Hotel booking with RAG FAQ"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # MCP SDK
    "mcp[cli]>=1.3.0",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.9",
    # Supabase — capped below 2.20.0 to avoid storage3>=2.28 pulling in pyiceberg
    # (pyiceberg requires MSVC C++ build tools on Windows to compile)
    "supabase>=2.10.0,<2.20.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    # Config
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    # LangChain / LangGraph
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-postgres>=0.0.12",
    # OpenAI
    "openai>=1.50.0",
    # Vector store
    "pgvector>=0.3.0",
    # Data
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    # Utilities
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
    "typer>=0.12.0",
    "aiofiles>=24.1.0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
cartrawler-server = "cartrawler.main:run_server"
cartrawler-seed = "scripts.seed_db:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cartrawler"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
